+++ src/torchjd/scalarization/__init__.py
@@ -25,5 +25,6 @@ from ._mean import Mean
 from ._random import Random
 from ._scalarizer_base import Scalarizer
 from ._sum import Sum
+from ._uw import UW
 
-__all__ = ["Constant", "GeometricMean", "Mean", "Random", "Scalarizer", "Sum"]
+__all__ = ["Constant", "GeometricMean", "Mean", "Random", "Scalarizer", "Sum", "UW"]